This is a cleanup not usually used.
If running regressiontests in a clean directory
then this is unnecessary (there is nothing to
save anyway), but usable there.

If you run regressiontests in the regressiontests source
(Do Not Do That) this might be useful, but you will have to
delete the few lines at the beginning that refuse to let this
run in the regressiontests source directory.
	modified:   CLEANUP.sh

+++ CLEANUP.sh
@@ -13,15 +13,28 @@ if test "x$hereb" = "xlibdwarf-regressiontests" ; then
   echo "Might be in regressiontest source here. Not cleaning up"
   exit 1
 fi
-(cd dwgena ; rm -f t*.base)
-(cd debugfission ; rm -f *.base)
+safequietrmf() {
+   if test -f $1 ; then
+     rm $1 2>/dev/null 1>/dev/null
+   fi
+}
+safequietrmd() {
+   if test -d $1 ; then
+     rmdir $1 2>/dev/null 1>/dev/null
+   fi
+}
+if test -d dwgena ; then
+  (cd dwgena ; rm -f t*.base)
+fi
+if test -d debugfission ; then
+  (cd debugfission ; rm -f *.base)
+fi
 # The following is copied from the kaufmann dir. Drop it.
 rm -f debugfissionb/t.o
 # The /tmp/dwa* files are lock files to prevent running two regr.
 # tests at the same time as two-at-once will not work!
 rm -f /tmp/dwa.*
 rm -f /tmp/dwb.*
-
 rm -f config.h.in~
 rm -f vglog
 rm -f sancheck.o
@@ -29,6 +42,7 @@ rm -f fuzz_*.o
 rm -f fuzz_* 
 rm -f dwgenc/runx
 rm -f dwgenc/testoutput.o
+safequietrmf implicitconst/t1.o
 rm -f filelist/filelist
 rm -f filelist/filelista
 rm -f filelist/filelistb
@@ -37,11 +51,19 @@ rm -f filelist/filelistd
 rm -f filelist/result
 rm -f filelist/localfuzz_init_path
 rm -f filelist/localfuzz_init_binary
+safequietrmf legendre/libmpich.so.1.0
+
 rm -f libdwallocs
 rm -f */libdwallocs
 rm -rf libbld
 rm -f libdwarf.a
 rm -f libdwarfp.a
+safequietrmf mustacchi/m32t.o
+safequietrmf mustacchi/m64t.o
+safequietrmf nolibelf/test.a
+safequietrmf nolibelf/test.o
+rm -f test-alex1/orig.a.out
+rm -f test-alex2/orig.a.out
 rm -f test_arange
 rm -f test_pubsreader
 rm -f test_bitoffset
@@ -54,6 +76,7 @@ rm -f dwn_t_out
 rm -f frame1/framexlocal.c
 rm -f frame1/frame2018.out
 rm -f frame1/selregs2018.out
+rm -f frame1/frame1
 rm -f tmp*
 rm -f testendian
 rm -f tmp*err*
@@ -162,5 +185,46 @@ rm -f base
 rm -f checkforlibzstd/a.out
 rm -f dwarfdump.O
 rm -f findfuncbypc
+# Intentionally we do not use rm -rf
+safequietrmd baddie1
+safequietrmd bigobj
+safequietrmd checkforlibz
+safequietrmd checkforlibzstd
+safequietrmd data16
+safequietrmd debugfission
+safequietrmd debugfissionb
+safequietrmd debuglink
+safequietrmd dwarfextract
+safequietrmf dwdebuglink
+safequietrmf  dwdiff.py
+safequietrmd dwgena
+safequietrmd dwgenc
+safequietrmf  dwnames_all
+safequietrmd emre2
+safequietrmd enciso4
+safequietrmf exfail.py 
+safequietrmd filelist
+safequietrmd findcu
+safequietrmd frame1
+safequietrmd guilfanov
+safequietrmd guilfanov2
+safequietrmd hughes2  
+safequietrmd implicitconst
+safequietrmd legendre
+safequietrmd moore
+safequietrmd  mustacchi
+safequietrmd  nolibelf
+safequietrmd  offsetfromlowpc
+safequietrmd  sandnes2
+safequietrmd  strsize
+safequietrmd  supplementary
+safequietrmd  test-alex1
+safequietrmd  test-alex2
+safequietrmd  testfindfuncbypc
+safequietrmd  testoffdie
+safequietrmf test_setframe
+safequietrmf test_simple_libfuncs
+safequietrmd  williamson
+
